pyproject.toml
[project]
name = "rps-game"
version = "0.1.0"
description = "剪刀石头布游戏 - Rock Paper Scissors Game for RK3588 ARM Linux"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "RPS Team"}
]
keywords = ["rock-paper-scissors", "gesture-recognition", "robotics", "rk3588"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",  # 使用默认上限，避免在Python 3.13下从源码构建旧版本
    "Pillow>=10.0.0",
    "mediapipe>=0.10.0",  # MediaPipe（代码已适配新旧版本）
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "types-PyYAML>=6.0.0",
]
serial = [
    "pyserial>=3.5",
]
all = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "types-PyYAML>=6.0.0",
    "pyserial>=3.5",
]

[project.scripts]
rps = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "types-PyYAML>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "mediapipe.*",
    "numpy.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.isort]
known-first-party = ["src"]
